amend rasa train CLI
