+++ rasa/cli/train.py
@@ -1,7 +1,6 @@
 import argparse
-import os
 import sys
-from typing import Dict, List, Optional, Text, TYPE_CHECKING, Union
+from typing import Dict, List, Optional, Text
 
 from rasa.cli import SubParsersAction
 import rasa.cli.arguments.train as train_arguments
@@ -9,18 +8,13 @@ import rasa.cli.arguments.train as train_arguments
 import rasa.cli.utils
 import rasa.utils.common
 from rasa.core.train import do_compare_training
-from rasa.shared.utils.cli import print_error
 from rasa.shared.constants import (
     CONFIG_MANDATORY_KEYS_CORE,
     CONFIG_MANDATORY_KEYS_NLU,
     CONFIG_MANDATORY_KEYS,
-    DEFAULT_CONFIG_PATH,
     DEFAULT_DOMAIN_PATH,
     DEFAULT_DATA_PATH,
 )
-
-if TYPE_CHECKING:
-    from pathlib import Path
 
 
 def add_subparser(
@@ -81,8 +75,10 @@ def run_training(args: argparse.Namespace, can_exit: bool = False) -> Optional[T
     domain = rasa.cli.utils.get_validated_path(
         args.domain, "domain", DEFAULT_DOMAIN_PATH, none_is_valid=True
     )
+    config = rasa.cli.utils.validate_config_path(args.config)
+    rasa.cli.utils.validate_assistant_id_in_config(config)
 
-    config = _get_valid_config(args.config, CONFIG_MANDATORY_KEYS)
+    config = rasa.cli.utils.get_valid_config(config, CONFIG_MANDATORY_KEYS)
 
     training_files = [
         rasa.cli.utils.get_validated_path(
@@ -146,7 +142,10 @@ def run_core_training(args: argparse.Namespace) -> Optional[Text]:
         if isinstance(args.config, list):
             args.config = args.config[0]
 
-        config = _get_valid_config(args.config, CONFIG_MANDATORY_KEYS_CORE)
+        config = rasa.cli.utils.validate_config_path(args.config)
+        rasa.cli.utils.validate_assistant_id_in_config(config)
+
+        config = rasa.cli.utils.get_valid_config(config, CONFIG_MANDATORY_KEYS_CORE)
 
         return train_core(
             domain=args.domain,
@@ -174,7 +173,10 @@ def run_nlu_training(args: argparse.Namespace) -> Optional[Text]:
     """
     from rasa.model_training import train_nlu
 
-    config = _get_valid_config(args.config, CONFIG_MANDATORY_KEYS_NLU)
+    config = rasa.cli.utils.validate_config_path(args.config)
+    rasa.cli.utils.validate_assistant_id_in_config(config)
+
+    config = rasa.cli.utils.get_valid_config(config, CONFIG_MANDATORY_KEYS_NLU)
     nlu_data = rasa.cli.utils.get_validated_path(
         args.nlu, "nlu", DEFAULT_DATA_PATH, none_is_valid=True
     )
@@ -215,41 +217,3 @@ def extract_nlu_additional_arguments(args: argparse.Namespace) -> Dict:
         arguments["num_threads"] = args.num_threads
 
     return arguments
-
-
-def _get_valid_config(
-    config: Optional[Union[Text, "Path"]],
-    mandatory_keys: List[Text],
-    default_config: Text = DEFAULT_CONFIG_PATH,
-) -> Text:
-    """Get a config from a config file and check if it is valid.
-
-    Exit if the config isn't valid.
-
-    Args:
-        config: Path to the config file.
-        mandatory_keys: The keys that have to be specified in the config file.
-        default_config: default config to use if the file at `config` doesn't exist.
-
-    Returns: The path to the config file if the config is valid.
-    """
-    config = rasa.cli.utils.get_validated_path(config, "config", default_config)
-
-    if not config or not os.path.exists(config):
-        print_error(
-            "The config file '{}' does not exist. Use '--config' to specify a "
-            "valid config file."
-            "".format(config)
-        )
-        sys.exit(1)
-
-    missing_keys = rasa.cli.utils.missing_config_keys(config, mandatory_keys)
-    if missing_keys:
-        print_error(
-            "The config file '{}' is missing mandatory parameters: "
-            "'{}'. Add missing parameters to config file and try again."
-            "".format(config, "', '".join(missing_keys))
-        )
-        sys.exit(1)
-
-    return str(config)
