+++ rasa/cli/utils.py
@@ -2,11 +2,20 @@ import json
 import logging
 import os
 import sys
+import time
 from types import FrameType
 from typing import Any, Dict, List, Optional, TYPE_CHECKING, Text, Union, overload
 
+import randomname
+
 import rasa.shared.utils.cli
 import rasa.shared.utils.io
+from rasa.shared.constants import (
+    ASSISTANT_ID_DEFAULT_VALUE,
+    ASSISTANT_ID_KEY,
+    DEFAULT_CONFIG_PATH,
+)
+from rasa.shared.utils.cli import print_error
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -99,6 +108,87 @@ def missing_config_keys(
     config_data = rasa.shared.utils.io.read_config_file(path)
 
     return [k for k in mandatory_keys if k not in config_data or config_data[k] is None]
+
+
+def validate_assistant_id_in_config(config_file: Union["Path", Text]) -> None:
+    """Verifies that the assistant_id key exists and has a unique value in config.
+
+    Issues a warning if the key does not exist or has the default value and replaces it
+    with a pseudo-random string value.
+    """
+    config_data = rasa.shared.utils.io.read_config_file(config_file)
+    assistant_id = config_data.get(ASSISTANT_ID_KEY)
+
+    if assistant_id is None or assistant_id == ASSISTANT_ID_DEFAULT_VALUE:
+        rasa.shared.utils.io.raise_warning(
+            f"The config file '{str(config_file)}' is missing a unique value for the "
+            f"'{ASSISTANT_ID_KEY}' mandatory key. Proceeding with generating a "
+            f"random value and overwriting the config file."
+        )
+
+        # add random value for assistant id, overwrite config file
+        time_format = "%Y%m%d-%H%M%S"
+        config_data[
+            ASSISTANT_ID_KEY
+        ] = f"{time.strftime(time_format)}-{randomname.get_name()}"
+
+        with open(config_file, "w") as config:
+            config.write(json.dumps(config_data))
+
+    return
+
+
+def validate_config_path(
+    config: Optional[Union[Text, "Path"]],
+    default_config: Text = DEFAULT_CONFIG_PATH,
+) -> Text:
+    """Verifies that the config path exists.
+
+    Exit if the config file does not exist.
+
+    Args:
+        config: Path to the config file.
+        default_config: default config to use if the file at `config` doesn't exist.
+
+    Returns: The path to the config file.
+    """
+    config = rasa.cli.utils.get_validated_path(config, "config", default_config)
+
+    if not config or not os.path.exists(config):
+        print_error(
+            "The config file '{}' does not exist. Use '--config' to specify a "
+            "valid config file."
+            "".format(config)
+        )
+        sys.exit(1)
+
+    return str(config)
+
+
+def get_valid_config(
+    config: Union[Text, "Path"],
+    mandatory_keys: List[Text],
+) -> Text:
+    """Get a config from a config file and check if it is valid.
+
+    Exit if the config isn't valid.
+
+    Args:
+        config: Path to the config file.
+        mandatory_keys: The keys that have to be specified in the config file.
+
+    Returns: The path to the config file if the config is valid.
+    """
+    missing_keys = rasa.cli.utils.missing_config_keys(config, mandatory_keys)
+    if missing_keys:
+        print_error(
+            "The config file '{}' is missing mandatory parameters: "
+            "'{}'. Add missing parameters to config file and try again."
+            "".format(config, "', '".join(missing_keys))
+        )
+        sys.exit(1)
+
+    return str(config)
 
 
 def cancel_cause_not_found(
