+++ rasa/shared/constants.py
@@ -63,7 +63,10 @@ TCP_PROTOCOL = "TCP"
 DEFAULT_SENDER_ID = "default"
 UTTER_PREFIX = "utter_"
 
-CONFIG_MANDATORY_COMMON_KEYS = ["assistant_id"]
+ASSISTANT_ID_KEY = "assistant_id"
+ASSISTANT_ID_DEFAULT_VALUE = "placeholder_default"
+
+CONFIG_MANDATORY_COMMON_KEYS = [ASSISTANT_ID_KEY]
 CONFIG_AUTOCONFIGURABLE_KEYS_CORE = ["policies"] + CONFIG_MANDATORY_COMMON_KEYS
 CONFIG_AUTOCONFIGURABLE_KEYS_NLU = ["pipeline"] + CONFIG_MANDATORY_COMMON_KEYS
 CONFIG_AUTOCONFIGURABLE_KEYS = (
