+++ tests/cli/test_rasa_train.py
@@ -1,5 +1,4 @@
 import os
-import tempfile
 import sys
 from pathlib import Path
 
@@ -17,12 +16,7 @@ from rasa.engine.storage.resource import Resource
 from rasa.shared.core.domain import Domain
 from rasa.model_training import CODE_NEEDS_TO_BE_RETRAINED, CODE_FORCED_TRAINING
 
-# noinspection PyProtectedMember
-from rasa.cli.train import _get_valid_config
 from rasa.shared.constants import (
-    CONFIG_MANDATORY_KEYS_CORE,
-    CONFIG_MANDATORY_KEYS_NLU,
-    CONFIG_MANDATORY_KEYS,
     LATEST_TRAINING_DATA_FORMAT_VERSION,
 )
 from rasa.shared.nlu.training_data.training_data import (
@@ -497,137 +491,6 @@ def test_train_core_help(run: Callable[..., RunResult]):
     printed_help = {line.strip() for line in output.outlines}
     for line in lines:
         assert line.strip() in printed_help
-
-
-@pytest.mark.parametrize(
-    "parameters",
-    [
-        {
-            "config_data": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-            },
-            "default_config": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-                "policies": ["TEDPolicy", "FallbackPolicy"],
-            },
-            "mandatory_keys": CONFIG_MANDATORY_KEYS_CORE,
-            "error": False,
-        },
-        {
-            "config_data": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-                "policies": None,
-            },
-            "default_config": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-                "policies": ["TEDPolicy", "FallbackPolicy"],
-            },
-            "mandatory_keys": CONFIG_MANDATORY_KEYS_CORE,
-            "error": False,
-        },
-        {
-            "config_data": {},
-            "default_config": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-                "policies": ["TEDPolicy", "FallbackPolicy"],
-            },
-            "mandatory_keys": CONFIG_MANDATORY_KEYS,
-            "error": True,
-        },
-        {
-            "config_data": {
-                "policies": ["TEDPolicy", "FallbackPolicy"],
-                "imports": "other-folder",
-            },
-            "default_config": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-                "policies": ["TEDPolicy", "FallbackPolicy"],
-            },
-            "mandatory_keys": CONFIG_MANDATORY_KEYS_NLU,
-            "error": True,
-        },
-        {
-            "config_data": None,
-            "default_config": {
-                "assistant_id": "placeholder_default",
-                "pipeline": "supervised",
-                "policies": ["TEDPolicy", "FallbackPolicy"],
-            },
-            "mandatory_keys": CONFIG_MANDATORY_KEYS_NLU,
-            "error": True,
-        },
-        {
-            "config_data": None,
-            "default_config": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-                "policies": ["TEDPolicy", "FallbackPolicy"],
-            },
-            "mandatory_keys": CONFIG_MANDATORY_KEYS,
-            "error": False,
-        },
-        {
-            "config_data": None,
-            "default_config": {
-                "assistant_id": "placeholder_default",
-                "language": "en",
-                "pipeline": "supervised",
-            },
-            "mandatory_keys": CONFIG_MANDATORY_KEYS_CORE,
-            "error": False,
-        },
-        {
-            "config_data": None,
-            "default_config": None,
-            "mandatory_keys": CONFIG_MANDATORY_KEYS,
-            "error": True,
-        },
-    ],
-)
-def test_get_valid_config(parameters):
-    config_path = None
-    if parameters["config_data"] is not None:
-        config_path = os.path.join(tempfile.mkdtemp(), "config.yml")
-        rasa.shared.utils.io.write_yaml(parameters["config_data"], config_path)
-
-    default_config_path = None
-    if parameters["default_config"] is not None:
-        default_config_path = os.path.join(tempfile.mkdtemp(), "default-config.yml")
-        rasa.shared.utils.io.write_yaml(
-            parameters["default_config"], default_config_path
-        )
-
-    if parameters["error"]:
-        with pytest.raises(SystemExit):
-            _get_valid_config(config_path, parameters["mandatory_keys"])
-
-    else:
-        config_path = _get_valid_config(
-            config_path, parameters["mandatory_keys"], default_config_path
-        )
-
-        config_data = rasa.shared.utils.io.read_yaml_file(config_path)
-
-        for k in parameters["mandatory_keys"]:
-            assert k in config_data
-
-
-def test_get_valid_config_with_non_existing_file():
-    with pytest.raises(SystemExit):
-        _get_valid_config("non-existing-file.yml", CONFIG_MANDATORY_KEYS)
 
 
 def test_train_nlu_finetune_with_model(
