+++ tests/cli/test_utils.py
@@ -8,6 +8,13 @@ import pytest
 from _pytest.logging import LogCaptureFixture
 
 import rasa.cli.utils
+from rasa.shared.constants import (
+    CONFIG_MANDATORY_KEYS,
+    CONFIG_MANDATORY_KEYS_CORE,
+    CONFIG_MANDATORY_KEYS_NLU,
+    DEFAULT_CONFIG_PATH,
+)
+import rasa.shared.utils.io
 from tests.cli.conftest import RASA_EXE
 
 
@@ -101,3 +108,144 @@ def test_validate_with_invalid_directory_if_default_is_valid(tmp_path: pathlib.P
         ) == str(tmp_path)
     assert len(record) == 1
     assert "does not seem to exist" in record[0].message.args[0]
+
+
+@pytest.mark.parametrize(
+    "parameters",
+    [
+        {
+            "config_data": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+            },
+            "default_config": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+                "policies": ["TEDPolicy", "FallbackPolicy"],
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS_CORE,
+            "error": False,
+        },
+        {
+            "config_data": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+                "policies": None,
+            },
+            "default_config": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+                "policies": ["TEDPolicy", "FallbackPolicy"],
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS_CORE,
+            "error": False,
+        },
+        {
+            "config_data": {},
+            "default_config": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+                "policies": ["TEDPolicy", "FallbackPolicy"],
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS,
+            "error": True,
+        },
+        {
+            "config_data": {
+                "policies": ["TEDPolicy", "FallbackPolicy"],
+                "imports": "other-folder",
+            },
+            "default_config": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+                "policies": ["TEDPolicy", "FallbackPolicy"],
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS_NLU,
+            "error": True,
+        },
+        {
+            "config_data": None,
+            "default_config": {
+                "assistant_id": "placeholder_default",
+                "pipeline": "supervised",
+                "policies": ["TEDPolicy", "FallbackPolicy"],
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS_NLU,
+            "error": True,
+        },
+        {
+            "config_data": None,
+            "default_config": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+                "policies": ["TEDPolicy", "FallbackPolicy"],
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS,
+            "error": False,
+        },
+        {
+            "config_data": None,
+            "default_config": {
+                "assistant_id": "placeholder_default",
+                "language": "en",
+                "pipeline": "supervised",
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS_CORE,
+            "error": False,
+        },
+        {
+            "config_data": None,
+            "default_config": None,
+            "mandatory_keys": CONFIG_MANDATORY_KEYS,
+            "error": True,
+        },
+    ],
+)
+def test_get_valid_config(parameters):
+    config_path = None
+
+    if parameters["config_data"] is not None:
+        config_path = os.path.join(tempfile.mkdtemp(), "config.yml")
+        rasa.shared.utils.io.write_yaml(parameters["config_data"], config_path)
+
+    default_config_path = os.path.join(tempfile.mkdtemp(), DEFAULT_CONFIG_PATH)
+
+    if parameters["default_config"] is not None:
+        default_config_path = os.path.join(tempfile.mkdtemp(), "default-config.yml")
+        rasa.shared.utils.io.write_yaml(
+            parameters["default_config"], default_config_path
+        )
+
+    if parameters["error"]:
+        with pytest.raises(SystemExit):
+            config_path = rasa.cli.utils.validate_config_path(
+                config_path, default_config_path
+            )
+            rasa.cli.utils.get_valid_config(config_path, parameters["mandatory_keys"])
+
+    else:
+        config_path = rasa.cli.utils.validate_config_path(
+            config_path, default_config_path
+        )
+
+        config_path = rasa.cli.utils.get_valid_config(
+            config_path,
+            parameters["mandatory_keys"],
+        )
+
+        config_data = rasa.shared.utils.io.read_yaml_file(config_path)
+
+        for k in parameters["mandatory_keys"]:
+            assert k in config_data
+
+
+def test_validate_config_path_with_non_existing_file():
+    with pytest.raises(SystemExit):
+        rasa.cli.utils.validate_config_path("non-existing-file.yml")
